Move trin launch into lib for easier testing
